move BA512Range to storage_blobs
