+++ sdk/core/src/errors.rs
@@ -146,15 +146,6 @@ pub enum HttpError {
     StreamResetError(StreamError),
 }
 
-/// An error caused by a range not being 512-byte aligned.
-#[derive(Debug, PartialEq, thiserror::Error)]
-pub enum Not512ByteAlignedError {
-    #[error("start range not 512-byte aligned: {0}")]
-    StartRange(u64),
-    #[error("end range not 512-byte aligned: {0}")]
-    EndRange(u64),
-}
-
 /// An error caused by invalid permissions.
 #[derive(Debug, thiserror::Error)]
 pub enum PermissionError {
@@ -165,17 +156,6 @@ pub enum PermissionError {
         received_token: char,
         supported_tokens: Vec<char>,
     },
-}
-
-/// An error caused by a range not being 512-byte aligned or by a parse failure.
-#[derive(Debug, PartialEq, thiserror::Error)]
-pub enum Parse512AlignedError {
-    #[error("split not found")]
-    SplitNotFound,
-    #[error("parse int error: {0}")]
-    ParseIntError(#[from] std::num::ParseIntError),
-    #[error("not 512 byte aligned error: {0}")]
-    Not512ByteAlignedError(#[from] Not512ByteAlignedError),
 }
 
 /// An error caused by failure to traverse a data structure.
