+++ sdk/storage_blobs/Cargo.toml
@@ -29,6 +29,7 @@ serde_json = "1.0"
 serde-xml-rs = "0.5"
 uuid = { version = "0.8", features = ["v4"] }
 url = "2.2"
+thiserror = "1.0"
 
 [dev-dependencies]
 tokio = { version = "1.0", features = ["full"] }
