+++ sdk/storage_blobs/examples/put_page_blob_00.rs
@@ -2,7 +2,7 @@
 extern crate log;
 use azure_core::prelude::*;
 use azure_storage::core::prelude::*;
-use azure_storage_blobs::prelude::*;
+use azure_storage_blobs::{prelude::*, BA512Range};
 use bytes::Bytes;
 use std::error::Error;
